requirements.txt
Flask==2.3.2
joblib==1.3.2
numpy==1.24.4
scikit-learn==1.3.2
snowflake-connector-python[pandas]==3.5.0
cryptography==41.0.5